requirements.txt
requests
feedparser
lxml
Pillow
datetime
beautifulsoup4
PyYAML
